refactor(edge): use same script form in all app
