--- app/edge/components/EdgeScriptForm/index.ts
@@ -1,1 +1,0 @@
-export { EdgeScriptForm, EdgeScriptFormAngular } from './EdgeScriptForm';
