--- app/edge/components/EdgeScriptForm/types.ts
@@ -1,11 +1,0 @@
-export type Platform = 'standalone' | 'swarm' | 'k8s';
-export type OS = 'win' | 'linux';
-
-export interface EdgeProperties {
-  os?: OS;
-  allowSelfSignedCertificates: boolean;
-  envVars: string;
-  edgeIdGenerator?: string;
-  platform?: Platform;
-  nomadToken?: string;
-}
