+++ app/edge/components/index.ts
@@ -1,9 +1,14 @@
 import angular from 'angular';
 
+import { r2a } from '@/react-tools/react2angular';
+import { EdgeScriptForm } from '@/react/edge/components/EdgeScriptForm';
+
 import { EdgeCheckinIntervalFieldAngular } from './EdgeCheckInIntervalField';
-import { EdgeScriptFormAngular } from './EdgeScriptForm';
 
 export const componentsModule = angular
   .module('app.edge.components', [])
-  .component('edgeCheckinIntervalField', EdgeCheckinIntervalFieldAngular)
-  .component('edgeScriptForm', EdgeScriptFormAngular).name;
+  .component(
+    'edgeScriptForm',
+    r2a(EdgeScriptForm, ['edgeInfo', 'commands', 'isNomadTokenVisible'])
+  )
+  .component('edgeCheckinIntervalField', EdgeCheckinIntervalFieldAngular).name;
