+++ app/react/edge/components/EdgeScriptForm/types.ts
@@ -1,0 +1,20 @@
+export type Platform = 'standalone' | 'swarm' | 'k8s' | 'nomad';
+export type OS = 'win' | 'linux';
+
+export interface ScriptFormValues {
+  nomadToken: string;
+  authEnabled: boolean;
+
+  allowSelfSignedCertificates: boolean;
+  envVars: string;
+
+  os: OS;
+  platform: Platform;
+
+  edgeIdGenerator?: string;
+}
+
+export interface EdgeInfo {
+  id?: string;
+  key: string;
+}
