+++ app/react/portainer/environments/wizard/EnvironmentsCreationView/WizardDocker/WizardDocker.tsx
@@ -5,7 +5,7 @@ import {
   BoxSelectorOption,
 } from '@/portainer/components/BoxSelector';
 import { Environment } from '@/portainer/environments/types';
-import { commandsTabs } from '@/edge/components/EdgeScriptForm/scripts';
+import { commandsTabs } from '@/react/edge/components/EdgeScriptForm/scripts';
 
 import { AnalyticsStateKey } from '../types';
 import { EdgeAgentTab } from '../shared/EdgeAgentTab';
