--- app/react/portainer/environments/wizard/EnvironmentsCreationView/shared/EdgeAgentTab/EdgeScriptForm/types.ts
@@ -1,12 +1,0 @@
-import { OS, Platform } from '@/edge/components/EdgeScriptForm/types';
-
-export interface ScriptFormValues {
-  nomadToken: string;
-  authEnabled: boolean;
-
-  allowSelfSignedCertificates: boolean;
-  envVars: string;
-
-  os: OS;
-  platform: Platform;
-}
