--- app/react/portainer/environments/wizard/EnvironmentsCreationView/shared/EdgeAgentTab/types.ts
@@ -1,4 +1,0 @@
-export interface EdgeInfo {
-  id?: string;
-  key: string;
-}
